requirement.txt
Django==2.1.8
pymysql==0.9.2
requests==2.18.4
djangorestframework==3.10.3
djangorestframework-filters==1.0.0.dev0
django-filters==0.2.1
django-cors-headers==3.2.0

clubfactory-thrift-client==2.1.23
schema==0.7.1
arrow==0.15.4
tornado==6.0.3